A few more optimizations.  It looks to be close to the end of the conversion.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3058 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/emucore/FrameBuffer.cxx
@@ -737,7 +737,7 @@ void FrameBuffer::setAvailableVidModes(uInt32 baseWidth, uInt32 baseHeight)
 
   for(auto& mode: myFullscreenModeLists)
     mode.clear();
-  for(int i = (int)myFullscreenModeLists.size(); i < myDisplays.size(); ++i)
+  for(size_t i = myFullscreenModeLists.size(); i < myDisplays.size(); ++i)
     myFullscreenModeLists.push_back(VideoModeList());
 
   // Check if zooming is allowed for this state (currently only allowed
